% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{get_etf_history}
\alias{get_etf_history}
\title{Get Any ETF History}
\usage{
get_etf_history(symbol, start_date, end_date, interval = "1day")
}
\arguments{
\item{symbol}{ETF ticker symbol}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{interval}{Data interval}
}
\value{
A data frame with historical data
}
\description{
Get Any ETF History
}
