% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{get_etf_price}
\alias{get_etf_price}
\title{Get Any ETF Price}
\usage{
get_etf_price(symbol)
}
\arguments{
\item{symbol}{ETF ticker symbol}
}
\value{
A list with price information
}
\description{
Get Any ETF Price
}
