% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{get_market_state}
\alias{get_market_state}
\title{Get Market State}
\usage{
get_market_state(exchange = "NYSE")
}
\arguments{
\item{exchange}{Exchange code (default: "NYSE")}
}
\value{
List with market state information
}
\description{
Get Market State
}
