% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gld_set_api_key}
\alias{gld_set_api_key}
\title{Set API key}
\usage{
gld_set_api_key(key)
}
\arguments{
\item{key}{Your Twelve Data API key}
}
\description{
Set API key
}
\examples{
\dontrun{
gld_set_api_key("your_api_key")
}
}
