% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{Options Functions (Requires Paid Subscription)}
\description{
Note: Options data from Twelve Data API requires a paid subscription.
These functions are placeholders and will return an error message.
}
\keyword{internal}
